Always use highp for color attributes

+++ src/shaders/debug.fragment.glsl
@@ -1,4 +1,4 @@
-uniform lowp vec4 u_color;
+uniform highp vec4 u_color;
 
 void main() {
     gl_FragColor = u_color;
